Add overly complicated cached embeddings generator
This adds a wrapper for embedding generator to cache embeddings for
identical documents. Might not really be useful in production, but was
fun to build and seems to work quite well! It's here almost 100% because
of my own curiosity and the fun challenge that was building it.

It's overly complex because all calls to Generate() are batches of
documents where none, some, or all may have been cached already. The
magic lies in using channels and a mutex when checking without blocking
other concurrent calls while also not making any duplicate calls (unless
they failed). For now, there's no TTL on cached embeddings, but that
could be added fairly easily if needed. Though honestly, I'm not sure
this cache is needed at all.